# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "contact": {
            "name": "Support Team",
            "email": "support@my-api.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/inventario": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Obtener lista de inventario",
                "description": "Retorna una lista del inventario",
                "operationId": "998be33b7a3fce13480669de4c864fd6",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "CodInv": {
                                                "description": "Código de inventario",
                                                "type": "integer"
                                            },
                                            "IdPrendaFk": {
                                                "description": "ID de prenda",
                                                "type": "integer"
                                            },
                                            "IdTallaFK": {
                                                "description": "ID de talla",
                                                "type": "integer"
                                            },
                                            "IdColorFK": {
                                                "description": "ID de color",
                                                "type": "integer"
                                            },
                                            "Cantidad": {
                                                "description": "Cantidad",
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Crear un nuevo item de inventario",
                "description": "Crea un nuevo item de inventario con los datos enviados",
                "operationId": "2131e4f4c8c6ccf3bbdf1a35eaa2fb6f",
                "requestBody": {
                    "description": "Información de inventario para crear un nuevo item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "CodInv": {
                                        "description": "Código de inventario",
                                        "type": "integer"
                                    },
                                    "IdPrendaFk": {
                                        "description": "ID de prenda",
                                        "type": "integer"
                                    },
                                    "IdTallaFK": {
                                        "description": "ID de talla",
                                        "type": "integer"
                                    },
                                    "IdColorFK": {
                                        "description": "ID de color",
                                        "type": "integer"
                                    },
                                    "Cantidad": {
                                        "description": "Cantidad",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item de inventario creado exitosamente"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
        "/api/inventario/{CodInv}": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Obtener un item de inventario",
                "description": "Retorna un item de inventario por su CodInv",
                "operationId": "bcee6c8bf1f1a0ae6fcb338d70ce09fe",
                "parameters": [
                    {
                        "name": "CodInv",
                        "in": "path",
                        "description": "Código de inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "CodInv": {
                                            "description": "Código de inventario",
                                            "type": "integer"
                                        },
                                        "IdPrendaFk": {
                                            "description": "ID de prenda",
                                            "type": "integer"
                                        },
                                        "IdTallaFK": {
                                            "description": "ID de talla",
                                            "type": "integer"
                                        },
                                        "IdColorFK": {
                                            "description": "ID de color",
                                            "type": "integer"
                                        },
                                        "Cantidad": {
                                            "description": "Cantidad",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Eliminar un item de inventario",
                "description": "Elimina un item de inventario por su CodInv",
                "operationId": "60b060902754f852997142b278efb42f",
                "parameters": [
                    {
                        "name": "CodInv",
                        "in": "path",
                        "description": "Código de inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item de inventario eliminado exitosamente"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
        "/api/inventario/{id}": {
            "put": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Actualizar un item de inventario",
                "description": "Actualiza un item de inventario con los datos enviados",
                "operationId": "de9eeb3a7306ade6b3c1fdf6d81d7687",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item de inventario a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Información de inventario para actualizar un item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "CodInv": {
                                        "description": "Código de inventario",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "IdPrendaFk": {
                                        "description": "ID de prenda",
                                        "type": "integer"
                                    },
                                    "IdTallaFK": {
                                        "description": "ID de talla",
                                        "type": "integer"
                                    },
                                    "IdColorFK": {
                                        "description": "ID de color",
                                        "type": "integer"
                                    },
                                    "Cantidad": {
                                        "description": "Cantidad",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item de inventario actualizado exitosamente"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
...
    },
...
}